[HU-06] refactor dtos and services location
