+++ web-smartstock/src/app/dtos/manufacturer.ts
