+++ web-smartstock/src/app/services/manufacturer.service.spec.ts
@@ -4,6 +4,7 @@ import { TestBed, async, inject } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ManufacturerService } from './manufacturer.service';
 
+
 describe('Service: Manufacturer', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
