+++ web-smartstock/src/app/services/manufacturer.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import { Manufacturer } from './manufacturer';
+import { Manufacturer } from '../dtos/manufacturer';
 
 @Injectable({
   providedIn: 'root'
